chore(deps): batch update 9 dependencies
Bumped:
- chrono-tz 0.9 → 0.10 (mneme-engine)
- page_size 0.4 → 0.6 (graph-builder)
- thiserror 1 → 2 (graph-builder)
- fastembed 4 → 5 (mneme)
- twox-hash 1.6 → 2.1 (mneme-engine)
- priority-queue 1.4 → 2.7 (mneme-engine)
- ndarray 0.15.6 → 0.16.1 (mneme-engine, mneme)
- ordered-float 4.2.0 → 5.1.0 (mneme-engine)
- rusqlite 0.33 → 0.38 (mneme, symbolon)

Breaking change fix: rusqlite 0.38 dropped usize: ToSql.
Changed get_history limit param from Option<usize> to Option<i64>.

Skipped: rayon pinned at =1.10.0 (graph_builder type conflict).

827 tests pass, 0 clippy warnings.

+++ crates/mneme/src/store.rs
@@ -263,7 +263,7 @@ impl SessionStore {
 
     /// Get message history for a session.
     #[instrument(skip(self))]
-    pub fn get_history(&self, session_id: &str, limit: Option<usize>) -> Result<Vec<Message>> {
+    pub fn get_history(&self, session_id: &str, limit: Option<i64>) -> Result<Vec<Message>> {
         let mut messages = Vec::new();
 
         if let Some(limit) = limit {
